Add files via upload
The APB Controller is a critical component in the AHB-APB bridge design, responsible for managing the communication between the AHB bus and the APB peripherals. The primary function of the APB Controller is to interface with the AHB Slave and generate the necessary control signals required for data transfers on the APB bus.