Add files via upload
The AHB Master is a central component in the AHB (Advanced High-performance Bus) system, responsible for initiating and controlling data transactions on the bus. As the master, it generates all necessary control signals and addresses, directing data traffic between memory, peripherals, and other bus components. The AHB Master plays a crucial role in ensuring efficient communication within the SoC (System-on-Chip) environment.


Uncomment the Appropriate Tasks for view the output